+++ packages/cli/src/commands/modalities/index.ts
@@ -1,0 +1,6 @@
+import { Command } from "commander";
+import { modalitiesGenerateCommand } from "./generate.js";
+
+export const modalitiesCommand = new Command("modalities")
+  .description("Work with NLX modalities")
+  .addCommand(modalitiesGenerateCommand);
